wallet/chain: reducing the max address in unspent api
Based on chain documentation at https://dashboard.chain.com/docs#bitcoin-address-unspents
you can send up to 200 addresses in the url. However maximum url limit
is  2048 and adviable length is 1024. 200 * 32 = 6400 . This limit is
not enforced strictly across internet but was causing a bug on the chain
side which would not show unconfirmed unspent when the number of
addresses exceeded around 120.

This commit changes the max addresses put into the url to 25 which
should give us 800 characters with plenty of room for our api keys and
the base path.

This change caused reduced performance for wallets with more than 25
addresses. (possibly by a factor 8 but
practically the reduced performance is around 1.5 based on
measurements on 500 address wallet). However, this gives us correct
results.

In future, we should try very hard to cache the api req going to chain.