tests/integration: Added integration fixtures to test two1
This change adds a couple fixtures that can be used to manipulate
the two1 CLI. It leverages the pexpect library.

Features:
* Fixture cli_runner and cli_runner_uninit to get a clean CLI to use.
They are initialized and non-initialized, respectively. Initialized means
that `two1 status` was run with a random username and location in tmp.
Therefore the wallet is created and the user created on the endpoint.
Uninit would require the user to create its own user, which would allow for tests
of initialization of the wallet or behavior of some commands when there is no
wallet or user config file
* Wallet sweeping. If a wallet was created (runner.walletCreated==True) then the
destructor of the runner will sweep the wallet to the preset SINK_HOLE/USER. This
is currently hardcoded to a @john user/wallet (that he can use for more testing!).
It cleans up both type of satoshis (on-chain and off-chain)

Run with:
```
python -u -m pytest -s two1/tests/integration/
```