cli/login: added the login command
This commit adds a new cli command: `21 login`.

21 login has two use cases:

    1- Account recovery: In case the user deletes their json file they
    can use 21 login to login into an old account
    2- Switch between accounts. Useful for demos and to go around multi
    user limitations in integ tests.

21 login hits an endpoint that is not under auth. Since it is possible
that a user without a username is calling 21 login we can authenticate
and sign the requests. However, the endpoint is protected with the
deviceid and public key. Whoever has both of these for a given user is
authorized to view all of the usernames created with that pair.