commands/config: add payment channel client to config object
This change allows us to refactor _get_balances to its original form,
and retrieve all balances without needing to pass a channel client
object. This also allows _get_balances to be more modular and
accessible in other circumstances (e.g. in `21 buy`).

This does add the overhead of having to load a channelclient any
time that a `Config` object is instantiated, but it is necessary
if we intend to re-use _get_balances.