status: show balances per username
Before this commit all the balances for a user were aggregated and shown
as total balance. This caused users to become confused when doing
bittransfers. Bittransfers are username based and although status showed
users have enough aggregated balance to buy, in reality they could not
buy due to their username not having enough balance.

With this commit, only the current users balance will be shown.