commands/login: simplify login logic
Prior to this commit, login was dependent upon lib.server.login
for the heavy logic of creating new users and signing users in.
All of that logic is now consolodated into the login command
script. The new changes make it impossible for a new user on a
bitcoin computer to create a new account using the `21 login`
commands. Instead `21 login` signs the new user into an existing
account as expected.