tests/integration: remove old style of integration tests
These tests have not been running for 3 months and are no longer stable.
A lot has changed in that time and it would be more work getting these
to work instead of writing new tests. There have been a lot of new unit
tests written so there is still good coverage.