bitrequests: Always use a MachineAuthWallet in BitTransferRequests
In many tutorials, BitTransferRequests is instantiated with a standard
Wallet object, instead of a MachineAuthWallet. This is preferred, since
all BitRequests types should implement a consistent constructor.

The API for signing messages is the same between MachineAuthWallet and
Wallet objects, but the methods for retrieving the wallet's signing
public key are different.

This change enforces use of a machine auth wallet, without affecting
the BitTransferRequests constructor or requiring any changes to
the 21.co/learn tutorials.