bitserv: Fix PaymentServer redeem to only update un-redeemed payments
This fixes the multi-machine redeem race by only redeeming payments
that have not yet been redeemed. This requires Flask/Django models
to return a boolean upon redeem, and support for both frameworks
is added.

Because no locks are needed or acquired, it reverses the order in
which a successful redeem will occur. With a multiprocess lock,
the first process to reach the top-level 'redeem' block will be
the one allowed to update the database. With the new pattern,
the first process to reach the database 'update' method will be
the one allowed. The tests are updated to reflect this change.