sell: Update message about 21 sell availability
As discussed w/ @balajis, this PR updates the message that is shown to
people when they try to run `21 sell` and see a message that the feature
is not available. We have had a few dozen people (40-50) reach out with
this issue since the 21 release. This commit provides clarity on what
distributions are available.

When other distributions become available, we can update the message.